requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.0.0
wordcloud>=1.8.0
tqdm>=4.64.0
pillow>=9.0.0
python-dotenv>=0.19.0

# YouTube/Instagram scraping
yt-dlp>=2023.0.0
gallery-dl>=1.20.0

# EXIF metadata
exiftool>=0.5.0

# Utilities
hurry.filesize>=0.9
requests>=2.28.0
beautifulsoup4>=4.11.0

# Development
black>=22.0.0
flake8>=5.0.0
pytest>=7.0.0 